feat: added subject variables to variable manager
As stated in issue #1396, variables added in message is recognized but variables added in subject is
not. So to fix this I went through the code understanding what was happening and found out that we
just needed to add subject string to the contents prop in component VariableManager. I hope it works
as I was unable to run then website on my system due to some problem after spending lot of my time
figuring it out.

#1396