update(activity-feed.md): show activity feed image
Changes Made:
Removed the link to the Activity Screen/Feed Image, and placed the image itself in its place. As the link is redirecting to a page that contains nothing but the image.

Previous:
A link was given to the image.

After Change:
The image itself is there in the place of the link.