Update README.md
Description:

The code of conduct paragraph was a bit long and confusing to read at first. Rewrote it to make it easier to read and understand. 

Reproduction Steps:
n/a

Expected Behavior:
n/a

Actual Behavior:
n/a

I have checked to see if this issue has already been raised and the pull requests have been closed.

I have read the contributing guidelines