Update apps/api/src/app/events/services/sms-service/handlers/bulksms.handler.ts
Hi Team, I have not created any provider package for bulksms I thought that it will be handled from your side. But if I am wrong please inform me so that I can make a package and will add it to novu. But for creating a package I just need your help.

Co-authored-by: George Djabarov <39195835+djabarovgeorge@users.noreply.github.com>