Add Pyroscope profiling service and relevant dependencies
This commit introduces a profiling service using Pyroscope in the `application-generic` package. Several environment variables for configuring the profiling service are added, and the relevant dependencies are included in the package.json file. The new service enables and initializes Pyroscope if it's enabled, and it is also exported in the index file of the package.