Add "@paralleldrive/cuid2" to API dependencies
This commit includes the addition of "@paralleldrive/cuid2" to API dependencies in the package.json file. The changes are reflected in the pnpm-lock.yaml file, demonstrating that the package has been successfully installed and is ready for use in the API's codebase. Additionally, the dependency 'debug' has been updated for several packages.