Add 'subscribers' field to WebhookResponseDto and update related methods
Added a new 'subscribers' field to the WebhookResponseDto and subsequently updated the convertToDTO function in the 'convert-response.usecase.ts' file to include this field. Additionally, updated the 'execute' method in 'get-webhook.usecase.ts' to return a Promise of 'WebhookResponseDto' type or null. Also, the method 'triggerWebhook' in 'webhook.controller.ts' has been significantly updated to handle webhook triggers more robustly.