Add index to optimize getMessages usecase
An index has been added to the message schema, specifically for the _environmentId and _subscriberId fields. This was done to optimize the execution of the getMessages usecase, where this specific combination of fields is being queried. It should improve the speed of fetching messages by ensuring the efficient retrieval of data.