fix: Ensure health check return the correct status
Currently, the DAL and Cache health checks won't cause
the overall health status of the app to return an error,
as Terminus requires us to throw a HealthCheck error
from the indicator when its status is unhealthy.

This PR throws the error when the indicator is unhealthy
and applies minor code refactorings to most health indicators.

For more information, refer to https://docs.nestjs.com/recipes/terminus#custom-health-indicator