Add Initial code owners for web and docker files (#5318)
* Create CODEOWNERS file and update .cspell.json

Introduced a new CODEOWNERS file to manage project ownership and review process in Github. Also, updated the .cspell.json with additional entries. This enhancement allows for a more structured code review process and improves the spell check functionality in the project.

* Update .cspell.json and remove unnecessary word entries

Several word entries such as 'Tomer' and 'devs' are now removed from .cspell.json file. This adjustment simplifies the spell check functionality in the project by limiting the dictionary to only necessary words.

* The provided commit message does not match the changes indicated in the diff section. The correct commit message would be:

Add ownership for Dockerfile to novu-infra team

The novu-infra team has been assigned as the owner of Dockerfile in all directories. This change ensures relevant notifications and permissions to the novu-infra team regarding modifications to Dockerfile.

* Update CODEOWNERS file

The "@novuhq/novu-devs" line is now commented out, effectively removing the novu-devs as default reviewers for most changes. This change was made pending team consensus.

---------

Co-authored-by: Richard Fontein <32132657+rifont@users.noreply.github.com>