ci(root): add feature flag checker (#5642)
* Add LaunchDarkly flag finder workflow

A new Github Actions workflow for finding LaunchDarkly feature flags was added to the codebase. Now we have better visibility regarding any feature flags that get introduced or removed with any PRs, helping maintain code quality. The spell check step has been temporarily disabled.

* Add team membership check to workflow

The commit adds a new step to the GitHub Actions workflow that checks if the creator of a pull request (PR) is a member of the specified team. If the PR creator is a team member, the workflow step for running the spell checker is skipped. This check is done using the GitHub API and the 'jq' command-line tool.