chore(root): Bump pnpm to its latest version (#5653)
* chore(root): Remove conventional commits

This is a small quality-of-life change for contributors. Since the introduction of the conventional commit PR title Github action and the introduction of squashed merges, there is no need to enforce conventional commits in our everyday commits before pushing them to the branch.

This will also speed up the husky hooks that take place locally.