Add step to set BullMQ token in pre-release job
In the pre-release job of the GitHub Actions Workflow, a step for setting the BullMQ token has been added. This token is then passed to the cloud deployment. We no longer use the secret directly in the deployment stage; instead, it's now set in the outputs of the pre-release job.