fix(web): Fetch current organization plan separately
Current organization plan (aka apiServiceLevel) is not migrated to Clerk for Novu EE. That is we can't do currentOrganization.apiServiceLevel anymore unless we pass it to the Clerk organization public metadata.

Since current plan need to stay in Mongo as well, we fetch it separately from a dedicated API endpoint.