fix(api,web): Align segment traits payload
Send the same payload during Segment Identify call from the API and the web app. This is a test to see if sometimes user data are lost on Segment's side when the traits sent from different apps differ.

It will be surprising to figure out that Segment doesn't merge traits based on user_id.