feat(application-generic): Add graceful shutdown module
Add a new module called "graceful-shutdown" to the "application-generic" library. This module includes a "ShutdownService" class that provides a timeout value for graceful shutdown. The timeout value is retrieved from the environment variable "GRACEFUL_SHUTDOWN_TIMEOUT" with a default value of 5000. This module also exports the "ShutdownService" for external use.